fixup! Fixed merge conflicts

+++ src/benchmark.h
@@ -13,7 +13,7 @@ struct sw_implementation
 {
 	const char* name;
 	int (*function)(const struct sequence_t*, const struct sequence_t*, const struct scores_t*);
-	long (*flops)(int, int);
+	size_t (*flops)(size_t, size_t);
 };
